Scheduler implemention

+++ src/main/java/com/project/UserPortal/Projection/ProjectWithCost.java
@@ -4,5 +4,5 @@ public interface ProjectWithCost
 {
     int getId();
     String getName();
-    double getCost();
+    Long getCost();
 }
